refactor: using dynamically dependencies CRDs
