+++ tests/e2e/k8s.go
@@ -466,7 +466,7 @@ func applyK8sResourceFromTemplate(kubectlOptions k8s.KubectlOptions, templateFil
 // extraArgs can be any kubectl api-resources parameter.
 func listK8sResourceKinds(kubectlOptions k8s.KubectlOptions, apiGroupSelector string, extraArgs ...string) ([]string, error) {
 	logMsg := "Listing K8s resource kind"
-	args := []string{"api-resources", "--verbs", "list", "-output", "name", "--sort-by", "name"}
+	args := []string{"api-resources", "--verbs", "list", "--output", "name", "--sort-by", "name"}
 
 	if apiGroupSelector != "" {
 		logMsg = fmt.Sprintf("%s group selector: %s", logMsg, apiGroupSelector)
